commit ejercicio 3_3

+++ src/Main.java
@@ -40,7 +40,7 @@ public class Main {
 
     // Calcula el subtotal de la compra
     public static double calcularSubtotal(List<String> productos, List<Double> precios, int[] cantidades) {
-        float subtotal = 0;
+        double subtotal = 0;
         for (int i = 0; i < productos.size(); i++) {
             // Error de control: verificar si la cantidad es mayor que cero
             subtotal += precios.get(i) * cantidades[i];
